ref(billing): Move quota_config_mapping into quota/ subdirectory
Relocate quota_config_mapping.py and its tests into a new
quota/ package under services/ for better organization.
